# dockstore-webservice/src/main/resources/swagger.yaml
...
swagger: "2.0"
info:
  description: "This describes the dockstore API, a webservice that manages pairs\
    \ of Docker images and associated metadata such as CWL documents and Dockerfiles\
    \ used to build those images"
  version: "1.8.2-SNAPSHOT"
  title: "Dockstore API"
  termsOfService: "TBD"
  contact:
    name: "Dockstore@ga4gh"
    url: "https://github.com/dockstore/dockstore"
    email: "theglobalalliance@genomicsandhealth.org"
  license:
    name: "Apache License Version 2.0"
    url: "https://github.com/dockstore/dockstore/blob/develop/LICENSE"
basePath: "/"
...
paths:
...
  /api/ga4gh/v2/extended/containers/{organization}:
    get:
      tags:
      - "extendedGA4GH"
      summary: "List entries of an organization"
      description: "This endpoint returns entries of an organization. "
      operationId: "entriesOrgGet"
      produces:
      - "application/json"
      - "text/plain"
      parameters:
      - name: "organization"
        in: "path"
        description: "An organization, for example `cancercollaboratory`"
        required: true
        type: "string"
      responses:
        200:
          description: "An array of Tools of the input organization."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ToolV1"
...
  /api/ga4gh/v2/extended/tools/entry/_search:
    post:
      tags:
      - "extendedGA4GH"
      summary: "Search the index of tools"
      description: "This endpoint searches the index for all published tools and workflows.\
        \ Used by utilities that expect to talk to an elastic search endpoint"
      operationId: "toolsIndexSearch"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "elastic search query"
        required: true
        schema:
          type: "string"
      responses:
        200:
          description: "An elastic search result."
          schema:
            type: "string"
  /api/ga4gh/v2/extended/tools/index:
    post:
      tags:
      - "extendedGA4GH"
      summary: "Update the index of tools"
      description: "This endpoint updates the index for all published tools and workflows. "
      operationId: "toolsIndexGet"
      produces:
      - "text/plain"
      parameters: []
      responses:
        200:
          description: "An array of Tools of the input organization."
      security:
      - BEARER: []
  /api/ga4gh/v2/extended/tools/{organization}:
    get:
      tags:
      - "extendedGA4GH"
      summary: "List tools of an organization"
      description: "This endpoint returns tools of an organization. "
      operationId: "toolsOrgGet"
      produces:
      - "application/json"
      - "text/plain"
      parameters:
      - name: "organization"
        in: "path"
        description: "An organization, for example `cancercollaboratory`"
        required: true
        type: "string"
      responses:
        200:
          description: "An array of Tools of the input organization."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ToolV1"
  /api/ga4gh/v2/extended/workflows/{organization}:
    get:
      tags:
      - "extendedGA4GH"
      summary: "List workflows of an organization"
      description: "This endpoint returns workflows of an organization. "
      operationId: "workflowsOrgGet"
      produces:
      - "application/json"
      - "text/plain"
      parameters:
      - name: "organization"
        in: "path"
        description: "An organization, for example `cancercollaboratory`"
        required: true
        type: "string"
      responses:
        200:
          description: "An array of Tools of the input organization."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ToolV1"
  /api/ga4gh/v2/extended/{id}/versions/{version_id}/{type}/tests/{relative_path}:
    post:
      tags:
      - "extendedGA4GH"
      summary: "Annotate test JSON with information on whether it ran successfully\
        \ on particular platforms plus metadata"
      description: "Test JSON can be annotated with whether they ran correctly keyed\
        \ by platform and associated with some metadata "
      operationId: "toolsIdVersionsVersionIdTypeTestsPost"
      produces:
      - "application/json"
      parameters:
      - name: "type"
        in: "path"
        description: "The type of the underlying descriptor. Allowable values include\
          \ \"CWL\", \"WDL\", \"NFL\"."
        required: true
        type: "string"
      - name: "id"
        in: "path"
        description: "A unique identifier of the tool, scoped to this registry, for\
          \ example `123456`"
        required: true
        type: "string"
      - name: "version_id"
        in: "path"
        description: "An identifier of the tool version for this particular tool registry,\
          \ for example `v1`"
        required: true
        type: "string"
      - name: "relative_path"
        in: "path"
        description: "A relative path to the test json as retrieved from the files\
          \ endpoint or the tests endpoint"
        required: true
        type: "string"
        pattern: ".+"
      - name: "platform"
        in: "query"
        description: "Platform to report on"
        required: true
        type: "string"
      - name: "platform_version"
        in: "query"
        description: "Version of the platform to report on"
        required: true
        type: "string"
      - name: "verified"
        in: "query"
        description: "Verification status, omit to delete key"
        required: false
        type: "boolean"
      - name: "metadata"
        in: "query"
        description: "Additional information on the verification (notes, explanation)"
        required: true
        type: "string"
      responses:
        200:
          description: "The tool test JSON response."
          schema:
            type: "object"
            additionalProperties:
              type: "object"
        401:
          description: "Credentials not provided or incorrect"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "The tool test cannot be found to annotate."
          schema:
            $ref: "#/definitions/Error"
      security:
      - BEARER: []
...
  /containers/{containerId}:
    get:
      tags:
      - "containers"
      summary: "Retrieve a tool."
      description: "This is one of the few endpoints that returns the user object\
        \ with populated properties (minus the userProfiles property)"
      operationId: "getContainer"
      produces:
      - "application/json"
      parameters:
      - name: "containerId"
        in: "path"
        description: "Tool ID"
        required: true
        type: "integer"
        format: "int64"
      - name: "include"
        in: "query"
        description: "Comma-delimited list of fields to include: validations"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/DockstoreTool"
      security:
      - BEARER: []
    put:
      tags:
      - "containers"
      summary: "Update the tool with the given tool."
      description: "Updates default descriptor paths, default Docker paths, default\
        \ test parameter paths, git url, and default version. Also updates tool maintainer\
        \ email, and private access for manual tools."
      operationId: "updateContainer"
      produces:
      - "application/json"
      parameters:
      - name: "containerId"
        in: "path"
        description: "Tool to modify."
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Tool with updated information"
        required: true
        schema:
          $ref: "#/definitions/DockstoreTool"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/DockstoreTool"
      security:
      - BEARER: []
    delete:
      tags:
      - "containers"
      summary: "Delete a tool."
      description: ""
      operationId: "deleteContainer"
      produces:
      - "application/json"
      parameters:
      - name: "containerId"
        in: "path"
        description: "Tool id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid "
      security:
      - BEARER: []
...
  /organizations:
    get:
      tags:
      - "organizations"
      summary: "List all available organizations."
      description: "NO Authentication"
      operationId: "getApprovedOrganizations"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Organization"
    post:
      tags:
      - "organizations"
      summary: "Create an organization."
      description: "Organization requires approval by an admin before being made public."
      operationId: "createOrganization"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Organization to register."
        required: true
        schema:
          $ref: "#/definitions/Organization"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Organization"
      security:
      - BEARER: []
...
  /workflows/{workflowId}:
    get:
      tags:
      - "workflows"
      summary: "Retrieve a workflow"
      description: "This is one of the few endpoints that returns the user object\
        \ with populated properties (minus the userProfiles property)"
      operationId: "getWorkflow"
      produces:
      - "application/json"
      parameters:
      - name: "workflowId"
        in: "path"
        description: "workflow ID"
        required: true
        type: "integer"
        format: "int64"
      - name: "include"
        in: "query"
        description: "Comma-delimited list of fields to include: validations, aliases"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Workflow"
      security:
      - BEARER: []
    put:
      tags:
      - "workflows"
      summary: "Update the workflow with the given workflow."
      description: "Updates descriptor type (if stub), default workflow path, default\
        \ file path, and default version"
      operationId: "updateWorkflow"
      produces:
      - "application/json"
      parameters:
      - name: "workflowId"
        in: "path"
        description: "Workflow to modify."
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Workflow with updated information"
        required: true
        schema:
          $ref: "#/definitions/Workflow"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Workflow"
      security:
      - BEARER: []
...
  /workflows/{workflowId}/tools/{workflowVersionId}:
    get:
      tags:
      - "workflows"
      summary: "Get the Tools for a given workflow version."
      description: "Does not require authentication for published workflows, authentication\
        \ can be provided for restricted workflows"
      operationId: "getTableToolContent"
      produces:
      - "application/json"
      parameters:
      - name: "workflowId"
        in: "path"
        description: "workflowId"
        required: true
        type: "integer"
        format: "int64"
      - name: "workflowVersionId"
        in: "path"
        description: "workflowVersionId"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
      security:
      - BEARER: []
...
securityDefinitions:
  BEARER:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  Alias:
    type: "object"
    properties:
      content:
        type: "string"
...
  DockstoreTool:
    type: "object"
    required:
    - "defaultCWLTestParameterFile"
    - "defaultWDLTestParameterFile"
    - "default_cwl_path"
    - "default_dockerfile_path"
    - "default_wdl_path"
    - "gitUrl"
    - "mode"
    - "name"
    - "namespace"
    - "private_access"
    - "registry_string"
    properties:
      aliases:
        type: "object"
        description: "aliases can be used as an alternate unique id for entries"
        additionalProperties:
          $ref: "#/definitions/Alias"
      custom_docker_registry_path:
        type: "string"
        readOnly: true
      dbCreateDate:
        type: "string"
        format: "date-time"
      dbUpdateDate:
        type: "string"
        format: "date-time"
      has_checker:
        type: "boolean"
        readOnly: true
      id:
        type: "integer"
        format: "int64"
        description: "Implementation specific ID for the container in this web service"
      input_file_formats:
        type: "array"
        readOnly: true
        uniqueItems: true
        items:
          $ref: "#/definitions/FileFormat"
      last_modified_date:
        type: "string"
        format: "date-time"
        readOnly: true
      output_file_formats:
        type: "array"
        readOnly: true
        uniqueItems: true
        items:
          $ref: "#/definitions/FileFormat"
      tags:
        type: "array"
        uniqueItems: true
        items:
          $ref: "#/definitions/Tag"
      topicId:
        type: "integer"
        format: "int64"
        description: "The Id of the corresponding topic on Dockstore Discuss"
      author:
        type: "string"
        position: 1
        description: "This is the name of the author stated in the Dockstore.cwl"
      description:
        type: "string"
        position: 2
        description: "This is a human-readable description of this container and what\
          \ it is trying to accomplish, required GA4GH"
      labels:
        type: "array"
        position: 3
        description: "Labels (i.e. meta tags) for describing the purpose and contents\
          \ of containers"
        uniqueItems: true
        items:
          $ref: "#/definitions/Label"
      users:
        type: "array"
        position: 4
        description: "This indicates the users that have control over this entry,\
          \ dockstore specific"
        uniqueItems: true
        items:
          $ref: "#/definitions/User"
      starredUsers:
        type: "array"
        position: 5
        description: "This indicates the users that have starred this entry, dockstore\
          \ specific"
        uniqueItems: true
        items:
          $ref: "#/definitions/User"
      email:
        type: "string"
        position: 6
        description: "This is the email of the git organization"
      defaultVersion:
        type: "string"
        position: 7
        description: "This is the default version of the entry"
      is_published:
        type: "boolean"
        position: 8
        description: "Implementation specific visibility in this web service"
      last_modified:
        type: "integer"
        format: "int32"
        position: 9
        description: "Implementation specific timestamp for last modified. Tools->\
          \ For automated/manual builds: N/A. For hosted: Last time a file was updated/created\
          \ (new version created). Workflows-> For remote: When refresh is hit, last\
          \ time GitHub repo was changed. Hosted: Last time a new version was made."
      lastUpdated:
        type: "string"
        format: "date-time"
        position: 10
        description: "Implementation specific timestamp for last updated on webservice.\
          \ Tools-> For automated builds: last time tool/namespace was refreshed Dockstore,\
          \ tool info (like changing dockerfile path) updated, or default version\
          \ selected. For hosted tools: when you created the tool. Workflows-> For\
          \ remote: When refresh all is hit for first time. Hosted: Seems to be time\
          \ created."
      gitUrl:
        type: "string"
        position: 11
        description: "This is a link to the associated repo with a descriptor, required\
          \ GA4GH"
      checker_id:
        type: "integer"
        format: "int64"
        position: 12
        description: "The id of the associated checker workflow"
        readOnly: true
      conceptDoi:
        type: "string"
        position: 13
        description: "The Digital Object Identifier (DOI) representing all of the\
          \ versions of your workflow"
      mode:
        type: "string"
        position: 13
        description: "This indicates what mode this is in which informs how we do\
          \ things like refresh, dockstore specific"
        enum:
        - "AUTO_DETECT_QUAY_TAGS_AUTOMATED_BUILDS"
        - "AUTO_DETECT_QUAY_TAGS_WITH_MIXED"
        - "MANUAL_IMAGE_PATH"
        - "HOSTED"
      name:
        type: "string"
        position: 14
        description: "This is the name of the container, required: GA4GH"
      default_dockerfile_path:
        type: "string"
        position: 15
        description: "This indicates for the associated git repository, the default\
          \ path to the Dockerfile, required: GA4GH"
      default_cwl_path:
        type: "string"
        position: 16
        description: "This indicates for the associated git repository, the default\
          \ path to the CWL document, required: GA4GH"
      default_wdl_path:
        type: "string"
        position: 17
        description: "This indicates for the associated git repository, the default\
          \ path to the WDL document"
      defaultCWLTestParameterFile:
        type: "string"
        position: 18
        description: "This indicates for the associated git repository, the default\
          \ path to the CWL test parameter file"
      defaultWDLTestParameterFile:
        type: "string"
        position: 19
        description: "This indicates for the associated git repository, the default\
          \ path to the WDL test parameter file"
      tool_maintainer_email:
        type: "string"
        position: 20
        description: "The email address of the tool maintainer. Required for private\
          \ repositories"
      private_access:
        type: "boolean"
        position: 21
        description: "Is the docker image private or not."
      toolname:
        type: "string"
        position: 22
        description: "This is the tool name of the container, when not-present this\
          \ will function just like 0.1 dockstorewhen present, this can be used to\
          \ distinguish between two containers based on the same image, but associated\
          \ with different CWL and Dockerfile documents. i.e. two containers with\
          \ the same registry+namespace+name but different toolnames will be two different\
          \ entries in the dockstore registry/namespace/name/tool, different options\
          \ to edit tags, and only the same insofar as they would \"docker pull\"\
          \ the same image, required: GA4GH"
      namespace:
        type: "string"
        position: 23
        description: "This is a docker namespace for the container, required: GA4GH"
      registry_string:
        type: "string"
        position: 24
        description: "This is a specific docker provider like quay.io or dockerhub\
          \ or n/a?, required: GA4GH"
      lastBuild:
        type: "string"
        format: "date-time"
        position: 25
        description: "Implementation specific timestamp for last built. For automated\
          \ builds: When refresh is hit, the last time the tool was built gets stored\
          \ here. If tool was never built on quay.io, then last build will be null.\
          \ N/A for hosted/manual path tools"
      workflowVersions:
        type: "array"
        position: 26
        description: "Implementation specific tracking of valid build tags for the\
          \ docker container"
        uniqueItems: true
        items:
          $ref: "#/definitions/Tag"
      path:
        type: "string"
        position: 27
      descriptorType:
        type: "array"
        position: 28
        readOnly: true
        items:
          type: "string"
      tool_path:
        type: "string"
        position: 29
        readOnly: true
      registry:
        type: "string"
        position: 30
        enum:
        - "QUAY_IO"
        - "DOCKER_HUB"
        - "GITLAB"
        - "AMAZON_ECR"
        - "SEVEN_BRIDGES"
    description: "This describes one entry in the dockstore. Logically, this currently\
      \ means one tuple of registry (either quay or docker hub), organization, image\
      \ name, and toolname which can be\n * associated with CWL and Dockerfile documents"
...
  Error:
    type: "object"
    required:
    - "code"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
...
  FileFormat:
    type: "object"
    required:
    - "value"
    properties:
      value:
        type: "string"
        position: 1
        description: "String representation of the file format"
    description: "This describes an input or output file format that is associated\
      \ with an entry in the dockstore"
...
  Organization:
    type: "object"
    required:
    - "name"
    - "status"
    - "users"
    properties:
      aliases:
        type: "object"
        description: "aliases can be used as an alternate unique id for organizations"
        additionalProperties:
          $ref: "#/definitions/Alias"
      dbCreateDate:
        type: "string"
        format: "date-time"
      dbUpdateDate:
        type: "string"
        format: "date-time"
      id:
        type: "integer"
        format: "int64"
        description: "Implementation specific ID for the organization in this web\
          \ service"
      name:
        type: "string"
        example: "OICR"
        position: 1
        description: "Name of the organization (ex. OICR)"
        minLength: 3
        maxLength: 39
        pattern: "[a-zA-Z][a-zA-Z\\d]*"
      description:
        type: "string"
        position: 2
        description: "Description of the organization"
      link:
        type: "string"
        position: 3
        description: "Link to the organization website"
      location:
        type: "string"
        position: 4
        description: "Location of the organization"
      email:
        type: "string"
        position: 5
        description: "Contact email for the organization"
      status:
        type: "string"
        position: 6
        description: "Is the organization approved, pending, or rejected"
        enum:
        - "PENDING"
        - "REJECTED"
        - "APPROVED"
      users:
        type: "array"
        position: 7
        description: "Set of users in the organization"
        uniqueItems: true
        items:
          $ref: "#/definitions/OrganizationUser"
      topic:
        type: "string"
        position: 8
        description: "Short description of the organization"
      avatarUrl:
        type: "string"
        position: 9
        description: "Logo URL"
        pattern: "([^\\s]+)(?i)(\\.jpg|\\.jpeg|\\.png|\\.gif)"
      displayName:
        type: "string"
        position: 9
        description: "Display name for an organization (Ex. Ontario Institute for\
          \ Cancer Research). Not used for links."
        minLength: 3
        maxLength: 50
        pattern: "[\\w ,_\\-&()']*"
      starredUsers:
        type: "array"
        position: 10
        description: "This indicates the users that have starred this organization"
        uniqueItems: true
        items:
          $ref: "#/definitions/User"
...
  ToolClass:
    type: "object"
    properties:
      description:
        type: "string"
        description: "A longer explanation of what this class is and what it can accomplish"
      id:
        type: "string"
        description: "The unique identifier for the class"
      name:
        type: "string"
        description: "A short friendly name for the class"
    description: "Describes a class (type) of tool allowing us to categorize workflows,\
      \ tasks, and maybe even other entities (such as services) separately"
...
  ToolV1:
    type: "object"
    properties:
      author:
        type: "string"
      contains:
        type: "array"
        items:
          type: "string"
      description:
        type: "string"
      id:
        type: "string"
      meta-version:
        type: "string"
      organization:
        type: "string"
      signed:
        type: "boolean"
      toolclass:
        $ref: "#/definitions/ToolClass"
      toolname:
        type: "string"
      url:
        type: "string"
      verified:
        type: "boolean"
      verified-source:
        type: "string"
      versions:
        type: "array"
        items:
          $ref: "#/definitions/ToolVersionV1"
    description: "A tool (or described tool) describes one pairing of a tool as described\
      \ in a descriptor file (which potentially describes multiple tools) and a Docker\
      \ image."
...
  ToolVersionV1:
    type: "object"
    properties:
      descriptor-type:
        type: "array"
        items:
          type: "string"
          enum:
          - "CWL"
          - "WDL"
      dockerfile:
        type: "boolean"
      id:
        type: "string"
      image:
        type: "string"
      meta-version:
        type: "string"
      name:
        type: "string"
      url:
        type: "string"
      verified:
        type: "boolean"
      verified-source:
        type: "string"
    description: "A tool version describes a particular iteration of a tool as described\
      \ by a reference to a specific image and dockerfile."
...